HC-4790 New API docs for mobile devs | Fix entity response
