+++ src/Response/ErrorResponse.php
@@ -2,15 +2,11 @@
 
 namespace Helpcrunch\Response;
 
-use Symfony\Component\HttpFoundation\JsonResponse;
-
-class ErrorResponse extends JsonResponse
+class ErrorResponse extends HelpcrunchResponse
 {
     const DEFAULT_ERROR_HEADERS = [
         'Access-Control-Allow-Origin' => '*'
     ];
-
-    protected static $defaultErrorMessageInfo = 'Something went wrong';
 
     public function __construct($message, string $innerErrorCode = null, int $status = self::HTTP_BAD_REQUEST, $errors = [])
     {
@@ -26,28 +22,5 @@ class ErrorResponse extends JsonResponse
         }
 
         parent::__construct($responseData, $status, self::DEFAULT_ERROR_HEADERS);
-    }
-
-    /**
-     * @param array|string $clientMessage
-     * @return array
-     */
-    private function getMessage($clientMessage): array
-    {
-        $message = ['info' => $this->getMessageInfo($clientMessage)];
-        if (is_array($clientMessage)) {
-            $message = array_merge($message, $clientMessage);
-        }
-
-        return $message;
-    }
-
-    /**
-     * @param array|string $clientMessage
-     * @return string
-     */
-    private function getMessageInfo($clientMessage): string
-    {
-        return is_string($clientMessage) ? $clientMessage : static::$defaultErrorMessageInfo;
     }
 }
